fix DRY violations
